@@ -1,5 +1,7 @@
 apply plugin: 'com.android.library'
 
+tasks.withType(Javadoc).all { enabled = false }
+
 ext {
     bintrayRepo = 'maven'
     bintrayName = 'stravazpot'
@@ -25,12 +27,12 @@ ext {
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 32
+    buildToolsVersion '27.3.1'
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 24
+        targetSdkVersion 32
         versionCode 7
         versionName "1.3.1"
     }
@@ -46,7 +48,7 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:25.3.1'
+    implementation 'androidx.appcompat:appcompat:1.0.2'   
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.squareup.okhttp3:okhttp:3.4.1'
     implementation 'com.squareup.retrofit2:retrofit:2.1.0'
